requirements.txt
python-dotenv==1.0.1
transformers==4.39.3
torch==2.2.2